Echo styling fixes
Given the conditional nature of the echo statements I thought it best to delete the echo factor out entirely. I opted instead for a dynamic variable that can change depending on the error the user encounters. The HTML should reference that variable in a paragraph tag that's styled to be at the bottom of the main div block.